Move common sheet action logic from cell view model to common view model
