+++ Wikipedia/Code/NotificationsCenterCellViewModel+SheetActionExtensions.swift
@@ -71,15 +71,15 @@ private extension NotificationsCenterCellViewModel {
     var userTalkPageActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
-        if let talkPageAction = titleTalkPageSheetAction(yourPhrasing: true) {
+        if let talkPageAction = commonViewModel.titleTalkPageAction(yourPhrasing: true) {
             sheetActions.append(talkPageAction)
         }
 
@@ -89,19 +89,19 @@ private extension NotificationsCenterCellViewModel {
     var mentionInTalkAndEditRevertedPageActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
-        if let titleTalkPageAction = titleTalkPageSheetAction(yourPhrasing: false) {
+        if let titleTalkPageAction = commonViewModel.titleTalkPageAction(yourPhrasing: false) {
             sheetActions.append(titleTalkPageAction)
         }
 
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             sheetActions.append(titleAction)
         }
 
@@ -111,15 +111,15 @@ private extension NotificationsCenterCellViewModel {
     var mentionInEditSummaryActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             sheetActions.append(titleAction)
         }
 
@@ -127,7 +127,7 @@ private extension NotificationsCenterCellViewModel {
     }
 
     var successfulAndFailedMentionActions: [NotificationsCenterAction] {
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             return [titleAction]
         }
 
@@ -137,15 +137,15 @@ private extension NotificationsCenterCellViewModel {
     var userGroupRightsActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let specificUserGroupRightsAction = specificUserGroupRightsSheetAction {
+        if let specificUserGroupRightsAction = commonViewModel.specificUserGroupRightsAction {
             sheetActions.append(specificUserGroupRightsAction)
         }
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let userGroupRightsAction = userGroupRightsSheetAction {
+        if let userGroupRightsAction = commonViewModel.userGroupRightsAction {
             sheetActions.append(userGroupRightsAction)
         }
 
@@ -155,11 +155,11 @@ private extension NotificationsCenterCellViewModel {
     var pageReviewedActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             sheetActions.append(titleAction)
         }
 
@@ -169,21 +169,21 @@ private extension NotificationsCenterCellViewModel {
     var pageLinkActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
         //Article where link was made
-        if let pageLinkToAction = pageLinkToAction {
+        if let pageLinkToAction = commonViewModel.pageLinkToAction {
             sheetActions.append(pageLinkToAction)
         }
         
         //Article you edited
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             sheetActions.append(titleAction)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
@@ -193,15 +193,15 @@ private extension NotificationsCenterCellViewModel {
     var connectionWithWikidataActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             sheetActions.append(titleAction)
         }
 
-        if let wikidataItemAction = wikidataItemAction {
+        if let wikidataItemAction = commonViewModel.wikidataItemAction {
             sheetActions.append(wikidataItemAction)
         }
 
@@ -209,7 +209,7 @@ private extension NotificationsCenterCellViewModel {
     }
 
     var emailFromOtherUserActions: [NotificationsCenterAction] {
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             return [agentUserPageAction]
         }
 
@@ -219,15 +219,15 @@ private extension NotificationsCenterCellViewModel {
     var thanksActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let titleAction = titleSheetAction {
+        if let titleAction = commonViewModel.titleAction {
             sheetActions.append(titleAction)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
@@ -237,11 +237,11 @@ private extension NotificationsCenterCellViewModel {
     var loginActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let loginHelpAction = loginNotificationsSheetAction {
+        if let loginHelpAction = commonViewModel.loginNotificationsAction {
             sheetActions.append(loginHelpAction)
         }
 
-        if let changePasswordSheetAction = changePasswordSheetAction {
+        if let changePasswordSheetAction = commonViewModel.changePasswordAction {
             sheetActions.append(changePasswordSheetAction)
         }
 
@@ -252,16 +252,16 @@ private extension NotificationsCenterCellViewModel {
         var sheetActions: [NotificationsCenterAction] = []
 
         if let secondaryLinks = notification.secondaryLinks {
-            let secondarySheetActions = secondaryLinks.compactMap { sheetActionForGenericLink(link:$0) }
+            let secondarySheetActions = secondaryLinks.compactMap { commonViewModel.actionForGenericLink(link:$0) }
             sheetActions.append(contentsOf: secondarySheetActions)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
         if let primaryLink = notification.primaryLink,
-           let primarySheetAction = sheetActionForGenericLink(link: primaryLink) {
+           let primarySheetAction = commonViewModel.actionForGenericLink(link: primaryLink) {
             sheetActions.append(primarySheetAction)
         }
 
@@ -271,165 +271,19 @@ private extension NotificationsCenterCellViewModel {
     var genericActions: [NotificationsCenterAction] {
         var sheetActions: [NotificationsCenterAction] = []
 
-        if let agentUserPageAction = agentUserPageSheetAction {
+        if let agentUserPageAction = commonViewModel.agentUserPageAction {
             sheetActions.append(agentUserPageAction)
         }
 
-        if let diffAction = diffSheetAction {
+        if let diffAction = commonViewModel.diffAction {
             sheetActions.append(diffAction)
         }
 
         if let primaryLink = notification.primaryLink,
-           let primarySheetAction = sheetActionForGenericLink(link: primaryLink) {
+           let primarySheetAction = commonViewModel.actionForGenericLink(link: primaryLink) {
             sheetActions.append(primarySheetAction)
         }
 
         return sheetActions
-    }
-}
-
-//MARK: Private Helpers - Individual Swipe Action methods
-
-private extension NotificationsCenterCellViewModel {
-    //Go to [Username]'s user page
-    var agentUserPageSheetAction: NotificationsCenterAction? {
-        guard let agentName = notification.agentName,
-              let url = customPrefixAgentNameURL(pageNamespace: .user) else {
-            return nil
-        }
-
-        let format = WMFLocalizedString("notifications-center-go-to-user-page", value: "Go to %1$@'s user page", comment: "Button text in Notifications Center that routes to a web view of the user page of the sender that triggered the notification. %1$@ is replaced with the sender's username.")
-        let text = String.localizedStringWithFormat(format, agentName)
-
-        let data = NotificationsCenterActionData(text: text, url: url)
-
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to diff
-    var diffSheetAction: NotificationsCenterAction? {
-        guard let url = fullTitleDiffURL else {
-            return nil
-        }
-
-        let text = WMFLocalizedString("notifications-center-go-to-diff", value: "Go to diff", comment: "Button text in Notifications Center that routes to a diff screen of the revision that triggered the notification.")
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to [your?] talk page
-    func titleTalkPageSheetAction(yourPhrasing: Bool = false) -> NotificationsCenterAction? {
-        
-        guard let linkData = linkData,
-              let namespace = linkData.titleNamespace,
-              let talkEquivalent = namespace.talkEquivalent,
-              let url = customPrefixTitleURL(pageNamespace: talkEquivalent) else {
-            return nil
-        }
-
-        let text = yourPhrasing ? WMFLocalizedString("notifications-center-go-to-your-talk-page", value: "Go to your talk page", comment: "Button text in Notifications Center that routes to user's talk page.") : WMFLocalizedString("notifications-center-go-to-talk-page", value: "Go to talk page", comment: "Button text in Notifications Center that routes to a talk page.")
-
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to [Name of article]
-    var titleSheetAction: NotificationsCenterAction? {
-        guard let linkData = linkData,
-              let url = fullTitleURL,
-              let title = notification.titleText else {
-            return nil
-        }
-        
-        var prefix = ""
-        if let namespace = linkData.titleNamespace {
-            prefix = namespace != .main ? "\(namespace.canonicalName):" : ""
-        }
-        let text = String.localizedStringWithFormat(CommonStrings.notificationsCenterGoToTitleFormat, "\(prefix)\(title)")
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to [Article where link was made]
-    var pageLinkToAction: NotificationsCenterAction? {
-        guard let url = pageLinkToURL,
-              let title = url.wmf_title else {
-            return nil
-        }
-
-        let text = String.localizedStringWithFormat(CommonStrings.notificationsCenterGoToTitleFormat, title)
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to Wikidata item
-    var wikidataItemAction: NotificationsCenterAction? {
-        guard let url = connectionWithWikidataItemURL else {
-            return nil
-        }
-
-        let text = WMFLocalizedString("notifications-center-go-to-wikidata-item", value: "Go to Wikidata item", comment: "Button text in Notifications Center that routes to a Wikidata item page.")
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to specific Special:UserGroupRights#{Type} page
-    var specificUserGroupRightsSheetAction: NotificationsCenterAction? {
-        guard let url = specificUserGroupRightsURL,
-              let type = url.fragment,
-              let title = url.wmf_title else {
-            return nil
-        }
-
-        let text = String.localizedStringWithFormat(CommonStrings.notificationsCenterGoToTitleFormat, "\(title)#\(type)")
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Go to Special:UserGroupRights
-    var userGroupRightsSheetAction: NotificationsCenterAction? {
-        guard let url = userGroupRightsURL,
-              let title = url.wmf_title else {
-            return nil
-        }
-
-        let text = String.localizedStringWithFormat(CommonStrings.notificationsCenterGoToTitleFormat, title)
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Login Notifications
-    var loginNotificationsSheetAction: NotificationsCenterAction? {
-        guard let url = loginNotificationsHelpURL else {
-            return nil
-        }
-
-        let text = WMFLocalizedString("notifications-center-login-notifications", value: "Login Notifications", comment: "Button text in Notifications Center that routes user to login notifications help page in web view.")
-
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    //Change password
-    var changePasswordSheetAction: NotificationsCenterAction? {
-
-        guard let url = changePasswordURL else {
-            return nil
-        }
-
-        let text = WMFLocalizedString("notifications-center-change-password", value: "Change Password", comment: "Button text in Notifications Center that routes user to change password screen.")
-
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
-    }
-
-    func sheetActionForGenericLink(link: RemoteNotificationLink) -> NotificationsCenterAction? {
-        guard let url = link.url,
-              let text = link.label else {
-            return nil
-        }
-
-        let data = NotificationsCenterActionData(text: text, url: url)
-        return NotificationsCenterAction.custom(data)
     }
 }
